pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-login>=0.6.3",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "qrcode>=8.2",
    "reportlab>=4.4.3",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.43",
    "pandas>=2.3.2",
    "openpyxl>=3.1.5",
    "xlwt>=1.3.0",
    "xlrd>=2.0.2",
    "python-escpos>=3.1",
    "pyusb>=1.3.1",
    "pyserial>=3.5",
    "charset-normalizer>=3.4.3",
]
